% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo}
\alias{demo}
\title{A sample Excel (.xlsx) export of SurveyMonkey response data.}
\format{
A data frame with 8 rows and 12 variables:
\describe{
...
}
}
\usage{
demo
}
\description{
A dataset containing the made up information about notable scientists.
}
\keyword{datasets}
