% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_mins.r
\name{flag_mins}
\alias{flag_mins}
\title{flag_mins}
\usage{
flag_mins(dat, mins)
}
\arguments{
\item{dat}{a data.frame containing SurveyMonkey response data, with two columns "date_created" and "date_modified".}

\item{mins}{a whole number specifying the cutoff for the .}
}
\value{
a data.frame object.
}
\description{
Flag survey duration based on how many minutes a respondent took. The assumption is that your data.frame contains two columns: "date_created" and "date_modified".
}
