% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_text.r
\name{flag_text}
\alias{flag_text}
\title{flag_text}
\usage{
flag_text(dat, text_vector, ignore_columns)
}
\arguments{
\item{dat}{a data.frame.}

\item{text_vector}{a character vector of values to identify in the data.frame.}

\item{ignore_columns}{a character vector of column names that will be ignored when flagging rows.}
}
\value{
a data.frame object.
}
\description{
Scan each row for elements in a character vector and create a boolean column flag_text if any element is present.

A use case would be to scan survey data for PID such as first/last names, scan free-text survey responses for foul language or non-alphanumeric values, etc.
}
