% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_cols.r
\name{trim_cols}
\alias{trim_cols}
\title{trim_cols}
\usage{
trim_cols(dat, ...)
}
\arguments{
\item{dat}{a data.frame}

\item{...}{a character vector of strings or partial strings to remove from column names}
}
\value{
a data.frame object
}
\description{
Trim column names by removing string matches or partial string matches.
}
