% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_headr.r
\name{trim_headr}
\alias{trim_headr}
\title{trim_headr}
\usage{
trim_headr(dat, ...)
}
\arguments{
\item{dat}{a data.frame}

\item{...}{a character vector of strings or partial strings to remove from column names}
}
\value{
a data.frame object
}
\description{
Trim column names by removing string matches or partial string matches.
}
