% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_headr.r
\name{trim_headr}
\alias{trim_headr}
\title{trim_headr}
\usage{
trim_headr(dat, ...)
}
\arguments{
\item{dat}{a data.frame.}

\item{...}{a character vector of values to remove from column names.}
}
\value{
a data.frame object.
}
\description{
Trim column names by removing string matches or partial string matches. Supply character string values to a character vector, and all matches will be removed from column names. For instance, supplying c('foo', 'bar') will remove all instances of 'foo' and/or 'bar' present in column names.
}
